Remove deprecated page resolvers
