+++ packages/hash/shared/src/queries/page.queries.ts
@@ -90,32 +90,3 @@ export const updatePersistedPageContents = gql`
 
   ${persistedPageFieldsFragment}
 `;
-
-const pagePropertiesFieldsFragment = gql`
-  fragment PagePropertyFields on PageProperties {
-    title
-    archived
-    icon
-  }
-`;
-
-export const updatePage = gql`
-  mutation updatePage(
-    $accountId: ID!
-    $entityId: ID!
-    $properties: PageUpdateData!
-  ) {
-    updatePage(
-      accountId: $accountId
-      entityId: $entityId
-      properties: $properties
-    ) {
-      accountId
-      entityId
-      properties {
-        ...PagePropertyFields
-      }
-    }
-  }
-  ${pagePropertiesFieldsFragment}
-`;
